demonio

+++ CG/src/estrutura/pontoZbufferConstante.java
@@ -13,7 +13,7 @@ import javafx.scene.paint.Color;
 public class pontoZbufferConstante {
     private Color cor = Color.WHITE;
     private double Z = Double.MIN_VALUE;
-    private double X = 0;
+    private int X = 0;
     
     public pontoZbufferConstante(){
     
@@ -24,7 +24,7 @@ public class pontoZbufferConstante {
         this.Z = Z;
     }
     
-    public pontoZbufferConstante(double X, double Z, int R, int G, int B){
+    public pontoZbufferConstante(int X, double Z, int R, int G, int B){
         this.cor = Color.rgb((int)R,(int)G,(int)B);
         this.Z = Z;
         this.X = X;
@@ -34,7 +34,7 @@ public class pontoZbufferConstante {
         return this.cor;
     }
     
-    public double getX(){
+    public int getX(){
         return this.X;
     }
